update bugs fix fat0010, fat00800, PMB01800, Pmt03500

+++ BIMASAKTI_11/1.00/PROGRAM/BS Program/SOURCE/COMMON/FA/FAT00100COMMON/DTOs/FAT0010002GetTransDetailResultDTO.cs
@@ -34,9 +34,9 @@ namespace FAT00100Common.DTOs
         public decimal NLCURRENCY_RATE { get; set; }
         public decimal NBBASE_RATE { get; set; }
         public decimal NBCURRENCY_RATE { get; set; }
-        public int NTOTAL_AMOUNT { get; set; }
-        public int NLTOTAL_AMOUNT { get; set; }
-        public int NBTOTAL_AMOUNT { get; set; }
+        public decimal NTOTAL_AMOUNT { get; set; }
+        public decimal NLTOTAL_AMOUNT { get; set; }
+        public decimal NBTOTAL_AMOUNT { get; set; }
         public string CCREATE_BY { get; set; } = string.Empty;
         public DateTime? DCREATE_DATE { get; set; }
         public string CUPDATE_BY { get; set; } = string.Empty;
